Added 'p1logger'
The p1logger script (which one should run in the background or as @reboot cron job) listens to the p1 port (via a P1 USB adapter) and logs the data into a .csv file. Currently, only 'Current power (W)' and 'Gas (m3)' will be logged, but extra options can be added easily according to the string-matching format (also see the .json file in this repo).